Move networkpolicies to a dedicated file.
