+++ pkg/operation/botanist/component/istio/istiod.go
@@ -17,7 +17,6 @@ package istio
 import (
 	"context"
 	"embed"
-	"fmt"
 	"path/filepath"
 	"time"
 
@@ -26,14 +25,11 @@ import (
 	"github.com/gardener/gardener/pkg/features"
 	gardenletfeatures "github.com/gardener/gardener/pkg/gardenlet/features"
 	"github.com/gardener/gardener/pkg/operation/botanist/component"
-	"github.com/gardener/gardener/pkg/operation/botanist/component/coredns"
-	"github.com/gardener/gardener/pkg/operation/botanist/component/nodelocaldns"
 	"github.com/gardener/gardener/pkg/utils/managedresources"
 
 	networkingv1alpha3 "istio.io/client-go/pkg/apis/networking/v1alpha3"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/intstr"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -271,241 +267,4 @@ func (i *istiod) generateIstiodChart() (*chartrenderer.RenderedChart, error) {
 		},
 		"image": i.values.Image,
 	})
-}
-
-
-type IstiodNetworkPolicyValues struct {
-	APIServerAddress string
-	// NodeLocalIPVSAddress is the CIDR of the node-local IPVS address.
-	NodeLocalIPVSAddress *string
-	// DNSServerAddress is the CIDR of the usual DNS server address.
-	DNSServerAddress *string
-}
-
-func getIstiodNetworkPolicyTransformers(values IstiodNetworkPolicyValues) []networkPolicyTransformer {
-	return []networkPolicyTransformer{
-		{
-			name: "deny-all",
-			transform: func(obj *networkingv1.NetworkPolicy) func() error {
-				return func() error {
-					obj.Annotations = map[string]string{
-						v1beta1constants.GardenerDescription: "Disables all Ingress and Egress traffic into/from this " +
-							"namespace.",
-					}
-					obj.Spec = networkingv1.NetworkPolicySpec{
-						PolicyTypes: []networkingv1.PolicyType{networkingv1.PolicyTypeEgress, networkingv1.PolicyTypeIngress},
-					}
-
-					return nil
-				}
-			},
-		},
-		{
-			name: "allow-to-seed-apiserver",
-			transform: func(obj *networkingv1.NetworkPolicy) func() error {
-				return func() error {
-					obj.Annotations = map[string]string{
-						v1beta1constants.GardenerDescription: "Allow egress to seed api-server.",
-					}
-					obj.Spec = networkingv1.NetworkPolicySpec{
-						PodSelector: metav1.LabelSelector{
-							MatchLabels: map[string]string{
-								v1beta1constants.LabelApp: "istiod",
-							},
-						},
-						Egress: []networkingv1.NetworkPolicyEgressRule{{
-							To: []networkingv1.NetworkPolicyPeer{
-								{
-									IPBlock: &networkingv1.IPBlock{CIDR: values.APIServerAddress}},	
-								},
-							Ports: []networkingv1.NetworkPolicyPort{
-									{Protocol: protocolPtr(corev1.ProtocolTCP), Port: intStrPtr(443)},
-								},
-							},
-						},
-						PolicyTypes: []networkingv1.PolicyType{networkingv1.PolicyTypeEgress},
-					}
-					return nil
-				}
-			},
-		},
-		{
-			name: "allow-to-dns",
-			transform: func(obj *networkingv1.NetworkPolicy) func() error {
-				return func() error {
-					obj.Annotations = map[string]string{
-						v1beta1constants.GardenerDescription: fmt.Sprintf("Allows Egress from pods labeled with "+
-							"'%s=%s' to DNS running in '%s'.", v1beta1constants.LabelApp, v1beta1constants.DefaultIngressGatewayAppLabelValue,
-							metav1.NamespaceSystem),
-					}
-					obj.Spec = networkingv1.NetworkPolicySpec{
-						PodSelector: metav1.LabelSelector{
-							MatchLabels: map[string]string{
-								v1beta1constants.LabelApp: v1beta1constants.DefaultIngressGatewayAppLabelValue,
-							},
-						},
-
-						Egress: []networkingv1.NetworkPolicyEgressRule{{
-							To: []networkingv1.NetworkPolicyPeer{
-								{
-									NamespaceSelector: &metav1.LabelSelector{
-										MatchLabels: map[string]string{
-											v1beta1constants.LabelRole: metav1.NamespaceSystem,
-										},
-									},
-									PodSelector: &metav1.LabelSelector{
-										MatchExpressions: []metav1.LabelSelectorRequirement{{
-											Key:      coredns.LabelKey,
-											Operator: metav1.LabelSelectorOpIn,
-											Values:   []string{coredns.LabelValue},
-										}},
-									},
-								},
-								{
-									NamespaceSelector: &metav1.LabelSelector{
-										MatchLabels: map[string]string{
-											v1beta1constants.LabelRole: metav1.NamespaceSystem,
-										},
-									},
-									PodSelector: &metav1.LabelSelector{
-										MatchExpressions: []metav1.LabelSelectorRequirement{{
-											Key:      coredns.LabelKey,
-											Operator: metav1.LabelSelectorOpIn,
-											Values:   []string{nodelocaldns.LabelValue},
-										}},
-									},
-								},
-							},
-							Ports: []networkingv1.NetworkPolicyPort{
-								{Protocol: protocolPtr(corev1.ProtocolUDP), Port: intStrPtr(coredns.PortServiceServer)},
-								{Protocol: protocolPtr(corev1.ProtocolTCP), Port: intStrPtr(coredns.PortServiceServer)},
-								{Protocol: protocolPtr(corev1.ProtocolUDP), Port: intStrPtr(coredns.PortServer)},
-								{Protocol: protocolPtr(corev1.ProtocolTCP), Port: intStrPtr(coredns.PortServer)},
-							},
-						}},
-						PolicyTypes: []networkingv1.PolicyType{networkingv1.PolicyTypeEgress},
-					}
-
-					if values.DNSServerAddress != nil {
-						obj.Spec.Egress[0].To = append(obj.Spec.Egress[0].To, networkingv1.NetworkPolicyPeer{
-							IPBlock: &networkingv1.IPBlock{
-								// required for node local dns feature, allows egress traffic to CoreDNS
-								CIDR: fmt.Sprintf("%s/32", *values.DNSServerAddress),
-							},
-						})
-					}
-
-					if values.NodeLocalIPVSAddress != nil {
-						obj.Spec.Egress[0].To = append(obj.Spec.Egress[0].To, networkingv1.NetworkPolicyPeer{
-							IPBlock: &networkingv1.IPBlock{
-								// required for node local dns feature, allows egress traffic to node local dns cache
-								CIDR: fmt.Sprintf("%s/32", *values.NodeLocalIPVSAddress),
-							},
-						})
-					}
-
-					return nil
-				}
-			},
-		},
-		{
-			name: "allow-from-vpn",
-			transform: func(obj *networkingv1.NetworkPolicy) func() error {
-				return func() error {
-					obj.Annotations = map[string]string{
-						v1beta1constants.GardenerDescription: "Allow ingress from vpn.",
-					}
-					obj.Spec = networkingv1.NetworkPolicySpec{
-						PodSelector: metav1.LabelSelector{
-							MatchLabels: map[string]string{
-								v1beta1constants.LabelApp: "istiod",
-							},
-						},
-						PolicyTypes: []networkingv1.PolicyType{networkingv1.PolicyTypeIngress},
-						Ingress: []networkingv1.NetworkPolicyIngressRule{{
-							From: []networkingv1.NetworkPolicyPeer{{
-								PodSelector: &metav1.LabelSelector{
-									MatchLabels: map[string]string{
-										v1beta1constants.LabelApp:   "vpn-shoot",
-										v1beta1constants.GardenRole: "system-component",
-									},
-								},
-								NamespaceSelector: &metav1.LabelSelector{},
-							}},
-							},
-						},
-					}
-					return nil
-				}
-			},
-		},
-		{
-			name: "allow-from-aggregate-prometheus",
-			transform: func(obj *networkingv1.NetworkPolicy) func() error {
-				return func() error {
-					obj.Annotations = map[string]string{
-						v1beta1constants.GardenerDescription: "Allow ingress from aggregate-prometheus.",
-					}
-					obj.Spec = networkingv1.NetworkPolicySpec{
-						PodSelector: metav1.LabelSelector{
-							MatchLabels: map[string]string{
-								v1beta1constants.LabelApp: "istiod",
-							},
-						},
-						PolicyTypes: []networkingv1.PolicyType{networkingv1.PolicyTypeIngress},
-						Ingress: []networkingv1.NetworkPolicyIngressRule{{
-							From: []networkingv1.NetworkPolicyPeer{{
-								PodSelector: &metav1.LabelSelector{
-									MatchLabels: map[string]string{
-										v1beta1constants.LabelApp:   "aggregate-prometheus",
-										v1beta1constants.GardenRole: "monitoring",
-									},
-								},
-								NamespaceSelector: &metav1.LabelSelector{},
-								}},
-								Ports: []networkingv1.NetworkPolicyPort{{
-									Protocol: protocolPtr(corev1.ProtocolTCP), 
-									Port:  func (port string) *intstr.IntOrString {v := intstr.FromString(port); return &v } ("metrics"),
-								}},
-							},
-						},
-					}
-					return nil
-				}
-			},
-		},
-		{
-			name: "allow-from-istio-ingress",
-			transform: func(obj *networkingv1.NetworkPolicy) func() error {
-				return func() error {
-					obj.Annotations = map[string]string{
-						v1beta1constants.GardenerDescription: "Allow ingress from istio-ingress.",
-					}
-					obj.Spec = networkingv1.NetworkPolicySpec{
-						PodSelector: metav1.LabelSelector{
-							MatchLabels: map[string]string{
-								v1beta1constants.LabelApp: "istiod",
-							},
-						},
-						PolicyTypes: []networkingv1.PolicyType{networkingv1.PolicyTypeIngress},
-						Ingress: []networkingv1.NetworkPolicyIngressRule{{
-							From: []networkingv1.NetworkPolicyPeer{{
-								PodSelector: &metav1.LabelSelector{
-									MatchLabels: map[string]string{
-										v1beta1constants.LabelApp:   "istio-ingressgateway",
-									},
-								},
-								NamespaceSelector: &metav1.LabelSelector{},
-							}},
-							// Ports: []networkingv1.NetworkPolicyPort{{
-							// 		Protocol: protocolPtr(corev1.ProtocolTCP), Port:  intstr.FromString("metrics")
-							// 	}},
-							},
-						},
-					}
-					return nil
-				}
-			},
-		},
-	}
 }
